ユーザーのエンドポイントにJob追加、フォル他構成一部変更
