+++ app/Domain/Entities/HtmlTag.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models\Domain\Entities;
+namespace App\Domain\Entities;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
