+++ app/Domain/Repositories/JobRepositoryInterface.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace App\Application\Contracts;
+namespace App\Domain\Repositories;
 
-use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Pagination\LengthAwarePaginator;
 
-interface JobUseCaseInterface
+interface JobRepositoryInterface
 {
     /**
      * 検索条件に基づいて求人情報を取得します。
@@ -15,7 +14,7 @@ interface JobUseCaseInterface
      * @param int|null $remoteId リモートID
      * @param array|null $tagIds タグIDの配列
      * @param int|null $limit 表示件数
-     * @return LengthAwarePaginator 全ての求人情報を含むLengthAwarePaginatorインスタンスを返します。
+     * @return LengthAwarePaginator
      * @throws ModelNotFoundException 検索条件に合致する求人情報が見つからない場合に投げられます。
      */
     public function getAllJobsBySearch(?int $categoryId, ?int $countryId, ?int $remoteId, ?array $tagIds, ?int $limit): LengthAwarePaginator;
@@ -23,8 +22,8 @@ interface JobUseCaseInterface
     /**
      * 指定されたIDの求人情報を取得します。
      *
-     * @param int $id 求人情報のID
-     * @return array 指定されたIDに対応する求人情報を連想配列として返します。
+     * @param int $id
+     * @return array
      * @throws ModelNotFoundException 指定されたIDに対応する求人情報が見つからない場合に投げられます。
      */
     public function getJobById(int $id): array;
