+++ app/Http/Requests/GetJobRequest.php
@@ -4,7 +4,7 @@ namespace App\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
-use App\Models\Domain\Entities\Job;
+use App\Domain\Entities\Job;
 
 class GetJobRequest extends FormRequest
 {
