+++ database/seeders/CountrySeeder.php
@@ -3,7 +3,7 @@
 namespace Database\Seeders;
 
 use Illuminate\Database\Seeder;
-use App\Models\Domain\Entities\Country;
+use App\Domain\Entities\Country;
 
 class CountrySeeder extends Seeder
 {
