+++ database/seeders/UserTagSeeder.php
@@ -3,9 +3,9 @@
 namespace Database\Seeders;
 
 use Illuminate\Database\Seeder;
-use App\Models\Domain\Entities\User;
-use App\Models\Domain\Entities\Tag;
-use App\Models\Domain\Entities\UserTag;
+use App\Domain\Entities\User;
+use App\Domain\Entities\Tag;
+use App\Domain\Entities\UserTag;
 
 class UserTagSeeder extends Seeder
 {
